prepare package creation
To have the crossword generator available in bigger javascript
projects it is helpful to add it to npmjs. This commit rewrites
the code and adds files required for uploading and installing a
package via npm.
CAUCTION: This commit will break the index.html. It is required
to use unpkg unce the node package is deployed.